Add debug print statement for group name in add_command function
In the add_command function, a debug print statement for the group name 'gname' has been added to aid in troubleshooting and monitoring. This print statement will output the group name to the console when a command is added. The addition of this print statement aims to improve the visibility of the group name variable during command addition operations.