Update git checkout command to use hash from search results
The previous implementation of the git checkout command was incorrectly using 'selected_text' instead of the actual hash value. This commit updates the subprocess.run(['git', 'checkout']) command to correctly use the hash value from the search results. This ensures that the correct commit is checked out when performing a search operation.